Refactor because editor only file in dir
