+++ src/lib/components/editor.svelte
